core-service/requirements.txt
grpcio
redis
protobuf